chore:  created the current user

+++ backend/src/services/recipes.js
@@ -1,13 +1,16 @@
 import { Recipe } from '../db/models/recipe.js'
 
-export async function createRecipe({
-  title,
-  author,
-  ingredientList,
-  imageURL,
-  tags,
-}) {
-  const recipe = new Recipe({ title, author, ingredientList, imageURL, tags })
+export async function createRecipe(
+  userId,
+  { title, ingredientList, imageURL, tags },
+) {
+  const recipe = new Recipe({
+    title,
+    author: userId,
+    ingredientList,
+    imageURL,
+    tags,
+  })
 
   return await recipe.save()
 }
@@ -43,17 +46,18 @@ export async function getRecipeById(recipeId) {
 
 // Update the recipe given a recipe ID ========================================
 export async function updateRecipe(
+  userId,
   recipeId,
-  { title, author, ingredientList, imageURL, tags },
+  { title, ingredientList, imageURL, tags },
 ) {
   return await Recipe.findOneAndUpdate(
-    { _id: recipeId },
-    { $set: { title, author, ingredientList, imageURL, tags } },
+    { _id: recipeId, author: userId },
+    { $set: { title, ingredientList, imageURL, tags } },
     { new: true },
   )
 }
 
 // Delete a post given a recipe ID ============================================
-export async function deleteRecipe(recipeId) {
-  return await Recipe.deleteOne({ _id: recipeId })
+export async function deleteRecipe(userId, recipeId) {
+  return await Recipe.deleteOne({ _id: recipeId, author: userId })
 }
